Moved from SystemIO.Compression to Archive
