+++ tests/OwlCore.Storage.Tests/Archive/Zip/IFileTests.cs
@@ -1,8 +1,8 @@
 ﻿using OwlCore.Storage.CommonTests;
-using OwlCore.Storage.SystemIO.Compression;
+using OwlCore.Storage.Archive;
 using System.IO.Compression;
 
-namespace OwlCore.Storage.Tests.SystemIO.Compression;
+namespace OwlCore.Storage.Tests.Archive.Zip;
 
 [TestClass]
 public class IFileTests : CommonIFileTests
@@ -27,7 +27,7 @@ public class IFileTests : CommonIFileTests
         var entry = archive.GetEntry(entryId);
         Assert.IsNotNull(entry);
 
-        var file = new ZipArchiveEntryFile(entry, new ZipArchiveFolder(archiveId, $"{Guid.NewGuid()}", archive));
+        var file = new ZipEntryFile(entry, new ZipFolder(archiveId, $"{Guid.NewGuid()}", archive));
 
         return Task.FromResult<IFile>(file);
