+++ tests/OwlCore.Storage.Tests/Archive/Zip/InMemIFileTests.cs
@@ -1,11 +1,11 @@
 ﻿using OwlCore.Storage.CommonTests;
-using OwlCore.Storage.SystemIO.Compression;
+using OwlCore.Storage.Archive;
 using System.IO.Compression;
 
-namespace OwlCore.Storage.Tests.SystemIO.Compression;
+namespace OwlCore.Storage.Tests.Archive.Zip;
 
 /// <summary>
-/// A test suite for <see cref="ZipArchiveEntryFile"/>s created entirely in memory.
+/// A test suite for <see cref="ZipEntryFile"/>s created entirely in memory.
 /// </summary>
 [TestClass]
 public class InMemIFileTests : CommonIFileTests
@@ -23,7 +23,7 @@ public class InMemIFileTests : CommonIFileTests
             entryStream.Write(randomData);
         }
 
-        var file = new ZipArchiveEntryFile(entry, new ZipArchiveFolder($"{Guid.NewGuid()}", $"{Guid.NewGuid()}", archive));
+        var file = new ZipEntryFile(entry, new ZipFolder($"{Guid.NewGuid()}", $"{Guid.NewGuid()}", archive));
 
         return Task.FromResult<IFile>(file);
